refactor: juniorPromiseRequest 파일 이동
