+++ src/pages/juniorPromiseRequest/components/Banner.tsx
