+++ src/pages/juniorPromiseRequest/utils/getDay.ts
