+++ src/pages/juniorPromiseRequest/utils/getSeniorValidWeekOfDay.ts
